+++ src/routes/transactions.ts
@@ -2,6 +2,11 @@ import { getTransactions } from '../models/transactionModel';
 import express from 'express';
 import { DynamoDBDocumentClient } from '@aws-sdk/lib-dynamodb';
 
+// TO DO hardcoded for now
+const USER_ID = "1";
+const BUDGET_ID = "1";
+const ACCT_ID = "1";
+
 export const router = express.Router();
 
 export default function (db: DynamoDBDocumentClient) {
@@ -17,18 +22,24 @@ export default function (db: DynamoDBDocumentClient) {
                 console.log("req.query.id is another type: " + typeof req.query.id);
             }
         }
-        const data = await getTransactions(db, transId);
+
+        const data = await getTransactions(db, {
+            userId: USER_ID,
+            budgetId: BUDGET_ID,
+            accountId: ACCT_ID,
+            transactionId: transId
+        });
+
         if (typeof data === 'string') {
             return res.json({msg: data, Count: 0});
         } else { // array of item objects
             res.json(data);
         }
-        
     });
 
-    // router.post('/', async (req, res) => {
-        
-    // });
+    router.post('/', async (req, res) => {
+        // read req.body
+    });
 
     return router;
 }
